Move content of static.py into counter.py
